grpc works without http gateway

+++ pkg/events/models/database_element.go
@@ -3,5 +3,5 @@ package models
 type DatabaseElement struct {
 	Type string
 	Count int32
-	Timestamp int64
+	Timestamp int32
 }
